Code cleanup and introducing login via http session

+++ src/main/java/com/pramurta/movie/controllers/AvailableShowController.java
@@ -1,10 +1,12 @@
 package com.pramurta.movie.controllers;
 
+import com.pramurta.movie.controllers.helpers.ControllerHelper;
 import com.pramurta.movie.domain.entities.AvailableShow;
 import com.pramurta.movie.domain.dtos.AvailableShowDto;
 import com.pramurta.movie.helpers.ResponseHelper;
 import com.pramurta.movie.mappers.AvailableShowMapper;
 import com.pramurta.movie.services.AvailableShowService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -28,9 +30,13 @@ public class AvailableShowController {
             @RequestParam(name = "movieName", required = false) String movieName,
             @RequestParam(name = "theatreName", required = false) String theatreName,
             @RequestParam(name = "fromTime", required = false) String fromTime,
-            @RequestParam(name = "toTime", required = false) String toTime
+            @RequestParam(name = "toTime", required = false) String toTime,
+            HttpSession httpSession
     ) {
         try {
+            if(!ControllerHelper.isUserLoggedIn(httpSession)) {
+                return new ResponseEntity<>(ControllerHelper.USER_NOT_LOGGED_IN_MESSAGE,HttpStatus.UNAUTHORIZED);
+            }
             List<AvailableShow> availableShows;
             if(movieName!=null && theatreName==null && fromTime==null && toTime==null) {
                 availableShows = availableShowService.getShowsByMovieName(movieName);
@@ -54,32 +60,44 @@ public class AvailableShowController {
                     .stream()
                     .map(availableShowMapper::mapToDto)
                     .toList();
-            return new ResponseEntity<>(ResponseHelper.constructSuccessfulAPIResponse(availableShowDtos), HttpStatus.OK);
+            return new ResponseEntity<>(ResponseHelper.constructSuccessResponse(availableShowDtos), HttpStatus.OK);
         } catch(Exception e) {
-            return new ResponseEntity<>(ResponseHelper.constructFailedAPIResponse(e.getMessage())
+            return new ResponseEntity<>(ResponseHelper.constructFailureResponse(e.getMessage())
                     ,HttpStatus.BAD_REQUEST);
         }
     }
 
     @PostMapping(path = "/availableShows")
-    public ResponseEntity<?> createAShow(@RequestBody AvailableShowDto availableShowDto) {
+    public ResponseEntity<?> createAShow(@RequestBody AvailableShowDto availableShowDto, HttpSession httpSession) {
         try {
+            if(!ControllerHelper.isUserLoggedIn(httpSession)) {
+                return new ResponseEntity<>("Please login first to create a show",HttpStatus.UNAUTHORIZED);
+            }
+            if(!ControllerHelper.isUserTheatreAdmin(httpSession)) {
+                return new ResponseEntity<>("Only theatre admins are allowed to create a show",HttpStatus.UNAUTHORIZED);
+            }
             AvailableShow availableShow = availableShowMapper.mapToEntity(availableShowDto);
             AvailableShowDto availableShowDtoCreated = availableShowMapper.mapToDto(availableShowService.createAShow(availableShow));
-            return new ResponseEntity<>(ResponseHelper.constructSuccessfulAPIResponse(availableShowDtoCreated),HttpStatus.CREATED);
+            return new ResponseEntity<>(ResponseHelper.constructSuccessResponse(availableShowDtoCreated),HttpStatus.CREATED);
         } catch(Exception e) {
-            return new ResponseEntity<>(ResponseHelper.constructFailedAPIResponse(e.getMessage())
-                    ,HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(ResponseHelper.constructFailureResponse(e.getMessage())
+                    ,HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping(path = "/availableShows/{showID}")
-    public ResponseEntity<?> removeAShow(@PathVariable("showID") String showID) {
+    public ResponseEntity<?> removeAShow(@PathVariable("showID") String showID, HttpSession httpSession) {
         try {
+            if(!ControllerHelper.isUserLoggedIn(httpSession)) {
+                return new ResponseEntity<>("Please login first to remove a show",HttpStatus.UNAUTHORIZED);
+            }
+            if(!ControllerHelper.isUserTheatreAdmin(httpSession)) {
+                return new ResponseEntity<>("Only theatre admins are allowed to create a show",HttpStatus.UNAUTHORIZED);
+            }
             availableShowService.removeAShow(showID);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
-            return new ResponseEntity<>(ResponseHelper.constructFailedAPIResponse(e.getMessage())
+            return new ResponseEntity<>(ResponseHelper.constructFailureResponse(e.getMessage())
                     ,HttpStatus.BAD_REQUEST);
         }
     }
